1. Merge remote dev at 375933475456407cdb47c455a9146f9aa93579c6.

2. Preserve the full u8 split range and stable two's-complement byte encoding.

3. Update Rholang evaluation and boundary checks to use the unsigned interface.

4. Retain the crypto golden vector and 129-send reduction regression.

5. Record the semantic merge audit and memory-bounded verification evidence.

+++ rholang/src/rust/interpreter/util/mod.rs
@@ -110,7 +110,7 @@ pub fn evaluation_random(
     } else if term_count > 256 {
         Ok(rand.split_short(index as u16 as i16))
     } else {
-        Ok(rand.split_byte(index as u8 as i8))
+        Ok(rand.split_byte(index as u8))
     }
 }
 
@@ -254,7 +254,7 @@ mod tests {
         for index in [0, 127, 128, 255] {
             assert_eq!(
                 evaluation_random(&random, index, 256).unwrap(),
-                random.split_byte(index as u8 as i8)
+                random.split_byte(index as u8)
             );
         }
     }
